Add import functionality for vmnameregistry_vmname resource; enhance documentation for VM name import process
