+++ provider/datasource_vmname.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -83,33 +82,4 @@ func dataSourceVmNameRead(d *schema.ResourceData, m interface{}) error {
 
 	d.SetId(vmName)
 	return nil
-}
-
-// Helper function to extract location from VM name
-func extractLocationFromVmName(vmName string) string {
-	// VM names follow pattern: lcp{env}{location}-{number}
-	// e.g., lcpdevuks-0001, lcpprduks-0002, etc.
-	parts := strings.Split(vmName, "-")
-	if len(parts) < 2 {
-		return ""
-	}
-
-	// Remove the "lcp" prefix and environment code to get location
-	prefix := parts[0]    // e.g., "lcpdevuks"
-	if len(prefix) <= 3 { // Should be at least "lcp" + something
-		return ""
-	}
-
-	// Remove "lcp" prefix
-	withoutLcp := prefix[3:] // e.g., "devuks"
-
-	// Common environment codes to remove
-	envCodes := []string{"dev", "prd", "ppd", "dvt"}
-	for _, envCode := range envCodes {
-		if strings.HasPrefix(withoutLcp, envCode) {
-			return withoutLcp[len(envCode):] // Return the location part
-		}
-	}
-
-	return ""
 }
